Folded ThreadExceptionDialog watchdog into HangDetection per review feedback
* Moved ThreadExceptionDialog detection/dismissal into HangDetection.WatchdogLoop so it runs during any InterruptWhenHung wait, not only during EndTest teardown
* Removed standalone ThreadExceptionDialogCanceler class
* Updated CloseOpenForm cross-reference to point at HangDetection.DismissThreadExceptionDialogs

See ai/todos/active/TODO-20260505_teardown_hang_fix.md
