--- pwiz_tools/Skyline/TestUtil/ThreadExceptionDialogCanceler.cs
@@ -1,161 +1,0 @@
-/*
- * Original author: Brian Pratt <bspratt .at. proteinms.net>,
- *                  MacCoss Lab, Department of Genome Sciences, UW
- *
- * Copyright 2026 University of Washington - Seattle, WA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading;
-using System.Windows.Forms;
-using pwiz.Common.SystemUtil;
-using pwiz.Skyline;
-using pwiz.Skyline.Util.Extensions;
-
-namespace pwiz.SkylineTestUtil
-{
-    /// <summary>
-    /// Background watchdog that dismisses any <see cref="ThreadExceptionDialog"/> that appears
-    /// during test teardown.
-    ///
-    /// If this watchdog ever fires it indicates a real bug: an exception escaped
-    /// <see cref="Application.ThreadException"/> (either an unhandled exception in test code,
-    /// or a framework race like the one this fix targets). The watchdog does not paper over
-    /// that bug - it converts a 30-minute UI-thread hang into a fast, loud test failure so a
-    /// developer has something actionable to investigate.
-    ///
-    /// Failure path: the captured dialog text is recorded via
-    /// <see cref="Program.AddTestException"/>, which is checked by
-    /// <c>AbstractFunctionalTest.RunFunctionalTest</c> at the end of the run and surfaced as
-    /// an <c>Assert.Fail</c> with the dialog body printed between separator lines.
-    ///
-    /// Why a ThreadExceptionDialog can appear during teardown: WinForms can catch an
-    /// exception inside a reentrant WndProc dispatch (e.g. Form.WmClose calling
-    /// <see cref="EventWaitHandle.Set"/> on a disposed SafeWaitHandle), and in that
-    /// reentrant path our registered <see cref="Application.ThreadException"/> handler is
-    /// sometimes bypassed and the default dialog appears, blocking the UI thread in a
-    /// nested message loop.
-    /// </summary>
-    public class ThreadExceptionDialogCanceler : IDisposable
-    {
-        private bool _disposed;
-
-        // Dialogs we have already logged and dispatched a dismiss for. Tracking by reference
-        // prevents the 500ms loop from flooding Program.TestExceptions and queueing redundant
-        // BeginInvoke callbacks if a dialog cannot be dismissed promptly (e.g. UI thread stuck
-        // in another nested loop). New dialogs that appear later are still handled.
-        private readonly HashSet<ThreadExceptionDialog> _handled = new HashSet<ThreadExceptionDialog>();
-
-        public ThreadExceptionDialogCanceler()
-        {
-            ActionUtil.RunAsync(ThreadProc, @"Thread Exception Dialog Canceler");
-        }
-
-        public void Dispose()
-        {
-            lock (this)
-            {
-                _disposed = true;
-            }
-        }
-
-        private void ThreadProc()
-        {
-            while (true)
-            {
-                lock (this)
-                {
-                    if (_disposed)
-                    {
-                        return;
-                    }
-                }
-
-                try
-                {
-                    DismissOpenDialogs();
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(@"ThreadExceptionDialogCanceler error: {0}", ex);
-                }
-
-                Thread.Sleep(500);
-            }
-        }
-
-        private void DismissOpenDialogs()
-        {
-            var dialogs = FormUtil.OpenForms.OfType<ThreadExceptionDialog>().ToList();
-            foreach (var dialog in dialogs)
-            {
-                if (dialog.IsDisposed || !dialog.IsHandleCreated)
-                {
-                    continue;
-                }
-
-                if (!_handled.Add(dialog))
-                {
-                    continue; // Already logged and dismissal scheduled for this dialog
-                }
-
-                Console.WriteLine(@"*** ThreadExceptionDialog detected during test teardown - dismissing");
-                Program.AddTestException(new InvalidOperationException(
-                    string.Format(@"ThreadExceptionDialog appeared during test teardown: {0}",
-                        TryGetDialogText(dialog))));
-
-                CommonActionUtil.SafeBeginInvoke(dialog, () =>
-                {
-                    try
-                    {
-                        // Setting DialogResult on a modal form posts WM_CLOSE asynchronously
-                        // (PostMessage), which avoids the synchronous reentrant Form.Close() path
-                        // that triggered the original hang.
-                        if (!dialog.IsDisposed)
-                        {
-                            dialog.DialogResult = DialogResult.Cancel;
-                        }
-                    }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine(@"Failed to dismiss ThreadExceptionDialog: {0}", ex);
-                    }
-                });
-            }
-        }
-
-        private static string TryGetDialogText(Form dialog)
-        {
-            // Best-effort diagnostic: this runs on a background thread, so any property access
-            // can throw if the dialog is disposed mid-call or cross-thread checks are enabled.
-            // Wrap the entire method so a throw here never breaks the watchdog loop.
-            try
-            {
-                // ThreadExceptionDialog's main message text is in a TextBox child control.
-                var textBox = dialog.Controls.OfType<TextBox>().FirstOrDefault(tb => tb.Multiline);
-                if (textBox != null && !string.IsNullOrEmpty(textBox.Text))
-                {
-                    return textBox.Text;
-                }
-                return dialog.Text ?? @"<no text>";
-            }
-            catch
-            {
-                return @"<unavailable>";
-            }
-        }
-    }
-}
